Reorganizar importaciones de AppLogoIcon para utilizar la ruta compartida
